Refactoring the performance feature
