+++ src/app/app.module.ts
@@ -89,12 +89,11 @@ import { CreateNewTestRunComponent } from './components/test/create-new-test-run
 import { UtilityComponent } from './components/utility/utility.component';
 import { SharedService } from './shared/core_apis/shared-utils';
 import { FileUploadModule } from 'primeng/fileupload';
-import { UploadFileComponent } from './components/utility/upload-file/upload-file.component';
-import { UtilityCategoryComponent } from './components/utility/utility-category.component';
-import { PerformanceTestsComponent } from './components/utility/performance-tests/performance-tests.component'; 
+import { UploadFileComponent } from './components/utility/utility-upload-file-tab/upload-file.component';
+import { UtilityPerformanceComponent } from './components/utility/utility-performance-tab/utility-performance.component';
+import { PerformanceTestsComponent } from './components/utility/performance-tests/performance-tests.component';
 import { PerformanceTestCasesListComponent } from './components/utility/performance-tests/performance-test-cases/performance-test-cases-list/performance-test-cases-list.component';
 import { PerformanceTestExecutionComponent } from './components/utility/performance-tests/performance-test-execution/performance-test-execution.component';
-import { PerformanceTestExecutionHistoryComponent } from './components/utility/performance-tests/performance-test-execution-history/performance-test-execution-history.component';
 import { PerformanceTestLogsComponent } from './components/utility/performance-tests/performance-test-execution/performance-test-logs/performance-test-logs.component';
 import { PerformanceTestLogToolbarComponent } from './components/utility/performance-tests/performance-test-execution/performance-test-logs/performance-test-log-toolbar/performance-test-log-toolbar.component';
 import { CreateNewPerformanceTestRunComponent } from './components/utility/performance-tests/create-new-performance-test-run/create-new-performance-test-run.component';
@@ -184,10 +183,9 @@ import { DndDirective } from './shared/dnd.directive';
     UtilityComponent,
     UploadFileComponent,
     PerformanceTestsComponent,
-    UtilityCategoryComponent,
+    UtilityPerformanceComponent,
     PerformanceTestCasesListComponent,
     PerformanceTestExecutionComponent,
-    PerformanceTestExecutionHistoryComponent,
     PerformanceTestLogsComponent,
     PerformanceTestLogToolbarComponent,
     CreateNewPerformanceTestRunComponent,
@@ -206,4 +204,3 @@ import { DndDirective } from './shared/dnd.directive';
     TestExecutionSandbox, WebSocketAPI, DialogService, DatePipe]
 })
 export class AppModule { }
-
