--- src/app/components/utility/performance-tests/performance-test-execution-history/performance-test-execution-history.component.scss
@@ -1,205 +1,0 @@
-
-/**
- *
- * Copyright (c) 2023 Project CHIP Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
- @import "../../../../../styles/colors";
-@import "../../../../../styles/custom.scss";
-:host ::ng-deep {
-  .p-datatable table {
-    border-collapse: collapse;
-    width: -webkit-fill-available;
-    table-layout: fixed;
-  }
-  .custom-times {
-    right: 1%;
-  }
-  .icon-add-circle1 {
-    font-size: 23px;
-    padding-left: 6px;
-    .path1:before {
-      color: $buttonColor;
-      opacity: 1;
-    }
-    .path2::before {
-      color: white;
-    }
-  }
-  .filter-button {
-    padding-right: 8px;
-  }
-  tr {
-    height: 60px;
-  }
-  .table_icons {
-    display: none;
-    i {
-      padding: 6px;
-      font-size: 25px;
-    }
-    i:hover {
-      color: $buttonColor;
-      background-color: rgba(156, 204, 235, 0.432);
-      border-radius: 20px;
-    }
-  }
-
-  .table_row:hover {
-    .table_icons {
-      display: block !important;
-    }
-  }
-  .p-paginator-bottom {
-    .p-dropdown {
-      width: 4.5rem !important;
-    }
-  }
-  .round_circle {
-    height: 10px;
-    width: 10px;
-
-    border-radius: 50%;
-    display: inline-block;
-  }
-  .failed {
-    background-color: $error;
-  }
-
-  .completed {
-    background-color: $success;
-  }
-  .pending {
-    background-color: $warning;
-  }
-  .table-heading-row th {
-    @extend .table-heading;
-  }
-  .table_row td {
-    @extend .table-data;
-  }
-  .p-paginator-bottom,
-  .p-paginator-page,
-  .p-dropdown-label,
-  .p-dropdown-items-wrapper,
-  .search-input {
-    @extend .table-data;
-  }
-  .project-dropdown {
-    .p-dropdown {
-      .p-inputtext {
-        @extend .heading-text;
-      }
-      .p-dropdown-trigger {
-        color: $iconBackground;
-      }
-      background: transparent !important;
-      box-shadow: unset;
-      border-color: transparent;
-    }
-    .p-dropdown:not(.p-disabled).p-focus {
-      box-shadow: unset;
-      border-color: transparent;
-      outline: unset;
-      outline-offset: unset;
-    }
-  }
-  .progress-bar .p-progressbar {
-    height: 6px;
-    width: 100px;
-    top: 10px;
-  }
-  .failed-p-bar {
-    .p-progressbar-value {
-      background-color: $error;
-    }
-  }
-  .completed-p-bar {
-    .p-progressbar-value {
-      background-image: linear-gradient(#007e33, #007e33),
-        linear-gradient(#cc0000, #cc0000), linear-gradient(#ffc933, #ffc933),
-        linear-gradient(#ff8800, #ff8800), linear-gradient(#2193f6, #2193f6),
-        linear-gradient(rgb(218, 209, 209), rgb(218, 209, 209));
-      background-repeat: no-repeat;
-      background-size: var(--percent-passed) 100%, var(--percent-failed) 100%,
-        var(--percent-error) 100%, var(--percent-cancelled) 100%,
-        var(--percent-pending) 100%, 100% 100%;
-    }
-  }
-  .pending-p-bar {
-    .p-progressbar-value {
-      background-color: $warning;
-    }
-  }
-  .test-name-td {
-    cursor: pointer;
-    &:hover {
-      color: $ceruleanBlue !important;
-    }
-  }
-  .legends-parent {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    .legends-title {
-      font-weight: 500;
-      padding-right: 10px;
-    }
-  }
-  .legends-color {
-    font-size: 15px;
-    .circle {
-      height: 8px;
-      width: 8px;
-      border-radius: 50%;
-      display: inline-block;
-    }
-    .pending {
-      background-color: #2193f6;
-    }
-    .error {
-      background-color: #ffc933;
-    }
-    .cancelled {
-      background-color: #ff8800;
-    }
-    .failed {
-      background-color: #cc0000;
-    }
-    .passed {
-      background-color: #007e33;
-    }
-  }
-  .spinner-height {
-    height: 100%;
-  }
-  .test-title {
-    width: 25%;
-    max-width: 0;
-    overflow: hidden;
-    white-space: pre;
-    text-overflow: ellipsis;
-  }
-  .import-loader-container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  
-  .import-loader-items {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-}
