+++ src/app/components/utility/performance-tests/performance-tests.component.html
@@ -43,8 +43,7 @@
       <div fxLayout fxFlex="1 1 auto" class="parent-test-cases box">
           <app-performance-test-suites-list fxFlex="40%"></app-performance-test-suites-list>
           <app-performance-test-cases-list fxFlex></app-performance-test-cases-list>
-
-        </div>
+      </div>
   </div>
 
   <div class="test-buttons" fxLayoutGap="10px" class="box">
