--- src/app/components/utility/utility-category.component.html
@@ -1,46 +1,0 @@
-<!--
- *
- * Copyright (c) 2023 Project CHIP Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
--->
-<div class="card-comp-parent-div">
-    <div class="p-card-parent">
-        <p-card class="p-card-comp">
-            <p-tabView #tabView (onChange)="tabViewChange($event)">
-                <!-- Tab 1 -->
-                <p-tabPanel header="File Upload">
-                    <app-upload-file></app-upload-file>
-                </p-tabPanel>
-
-                <!-- Tab 2 -->
-                <p-tabPanel header="Stress and Stability Test">
-                    <ng-template pTemplate="content"> 
-                        <ng-container *ngIf="sharedAPI.getSelectedProjectType()">
-                            <app-create-new-performance-test-run
-                                *ngIf="this.sharedAPI.getIsPerformanceTypeSelected()===0 && this.testSandBox.getPerformanceTestScreen() ===0">
-                            </app-create-new-performance-test-run>
-                            <app-performance-tests
-                                *ngIf="this.sharedAPI.getIsPerformanceTypeSelected()==1 && this.testSandBox.getPerformanceTestScreen() ===0">
-                            </app-performance-tests>
-                            <app-performance-test-execution
-                                *ngIf="this.testSandBox.getPerformanceTestScreen() ===1"></app-performance-test-execution>
-                            <app-performance-test-execution-history
-                                *ngIf="this.testSandBox.getPerformanceTestScreen() ===2"></app-performance-test-execution-history>
-                        </ng-container>
-                    </ng-template>
-                </p-tabPanel>
-            </p-tabView>
-        </p-card>
-    </div>
-</div>
