--- src/app/components/utility/utility-category.component.scss
@@ -1,230 +1,0 @@
-/**
- *
- * Copyright (c) 2023 Project CHIP Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@import "../../../styles/colors";
-@import "../../../styles/custom.scss";
-
-:host ::ng-deep {
-  .search-test-case {
-    position: absolute;
-    right: 0;
-    top: 53px;
-
-    input {
-      width: 11vw;
-    }
-  }
-  .test-name {
-    display: flex;
-    flex-direction: column;
-    gap: 1px;
-    height: calc(100vh - 495px);
-    overflow-y: auto;
-    border: solid 1px #dee2e6;
-    background-color: #fff;
-    position: relative;
-    padding-bottom: 10px;
-  }
-
-  .test-name-master {
-    background-color: #f8f9fa;
-    height: 45px;
-    display: flex;
-    border: solid 1px #dee2e6;
-
-    .master-text {
-      margin-bottom: 15px;
-    }
-  }
-
-  .test-name-checkbox {
-    margin: 15px 15px 0;
-  }
-
-  .master-label {
-    margin-top: 14px;
-    font-weight: bold;
-    font-size: 14px;
-  }
-
-  .custom-icon {
-    .p-checkbox .p-checkbox-box {
-      border-color: #2196f3;
-      background: #2196f3;
-
-      .p-checkbox-icon {
-        margin-bottom: 4px;
-        font-size: larger;
-      }
-
-      .p-checkbox-icon::before {
-        content: "\2212";
-        font-weight: bolder;
-      }
-    }
-  }
-
-  .label-text {
-    font-size: 14px;
-  }
-
-  .test-case-view {
-    height: 100%;
-
-    .sub-heading {
-      margin: 30px 1px 0 1px;
-    }
-
-    .scroll-test {
-      color: black;
-      background: #238f83;
-      /* width: fit-content; */
-      /* height: 26px; */
-      padding: 11px 0;
-      position: absolute;
-
-    }
-
-    .pi-angle-right {
-      right: 23%;
-      top: 62px;
-      z-index: 9999;
-    }
-
-    .pi-angle-left {
-      position: absolute;
-      top: 64px;
-      z-index: 9999;
-    }
-  }
-
-  .testcase-custom-view {
-    font-size: 14px;
-
-    ul {
-      width: 77%;
-      /* line-break: anywhere; */
-      /* white-space: nowrap; */
-      display: flex;
-      flex-direction: row;
-      flex-wrap: nowrap;
-      overflow-x: auto;
-      height: 58px;
-      border: none;
-    }
-
-    .parent-test-cases {
-      ::-webkit-scrollbar {
-        display: block !important;
-      }
-    }
-
-    ::-webkit-scrollbar {
-      display: none;
-    }
-  }
-
-  .testcase-custom-view,
-  .p-tabview-panel,
-  .p-tabview,
-  .parent-test-cases {
-    height: 100%;
-  }
-  .p-tabview-panels {
-    padding: 0 !important;
-    height: 94%;
-  }
-
-  .isDisabled {
-    pointer-events: none;
-    opacity: 0.59;
-  }
-
-  .back {
-    background-color: #FF5733;
-  }
-  .card-comp-parent-div {
-    height: calc(100vh - 97px);
-    width: calc(100vw - 114px);
-    margin: 14px;
-    .p-card-parent {
-      height: 100%;
-    }
-  }
-  .p-card-comp {
-    flex: 1 1 auto;
-    & > .p-component {
-      height: 100%;
-      box-shadow: 0px 6px 12px 3px rgb(0 0 0 / 10%);
-    }
-    .sub-heading {
-      font-size: 18px;
-    }
-    /* Override CSS */
-    .p-card-body {
-      padding: 30px;
-    }
-    .p-card-content {
-      padding: 0;
-    }
-    .heading-text {
-      @extend .heading-text;
-    }
-    .sub-heading {
-      font-family: "Roboto-Regular";
-      font-size: 16px;
-      color: $lightBlack;
-    }
-    .textbox-label {
-      font-family: "Roboto-Regular";
-      font-size: 14px;
-      color: $bluishGray;
-    }
-    h3 {
-      padding: 0;
-      margin: 0;
-    }
-    .p-card-body {
-      height: 100%;
-      padding: 0;
-      .p-card-content {
-        height: 100%;
-      }
-    }
-    .test-parent,
-    .app-test-component {
-      height: 100%;
-    }
-    .test-parent {
-      padding: 15px;
-    }
-    p-badge .p-badge.p-component {
-      min-width: 20px;
-      height: 20px;
-      line-height: 20px;
-    }
-    .success {
-      color: $success;
-    }
-    .error {
-      color: $error;
-    }
-    .pi-spinner {
-      color: $warning;
-    }
-  }
-}
