+++ src/app/components/utility/utility-performance-tab/utility-performance.component.html
@@ -1,0 +1,30 @@
+<!--
+ *
+ * Copyright (c) 2023 Project CHIP Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+-->
+<ng-container *ngIf="sharedAPI.getSelectedProjectType()">
+    <app-create-new-performance-test-run
+        *ngIf="sharedAPI.getIsPerformanceTypeSelected()===0 && testSandBox.getPerformanceTestScreen() ===0">
+    </app-create-new-performance-test-run>
+    <app-performance-tests
+        *ngIf="sharedAPI.getIsPerformanceTypeSelected()==1 && testSandBox.getPerformanceTestScreen() ===0">
+    </app-performance-tests>
+    <app-performance-test-execution
+        *ngIf="testSandBox.getPerformanceTestScreen() ===1">
+    </app-performance-test-execution>
+    <app-test-execution-history
+        *ngIf="testSandBox.getPerformanceTestScreen() ===2">
+    </app-test-execution-history>
+</ng-container>
