+++ src/app/components/utility/utility-upload-file-tab/upload-file.component.scss
