+++ src/app/components/utility/utility.component.spec.ts
@@ -47,13 +47,4 @@ describe('UtilityComponent', () => {
     sharedAPI = TestBed.inject(SharedAPI);
     date = TestBed.inject(DatePipe);
   });
-
-  it('should return difference in time', () => {
-    expect(component.getTimeDifference(0, 0)).toBe('-');
-  });
-
-  it('should return True or False', () => {
-    expect(component.isFieldDisplayed(1)).toBe(true);
-    expect(component.isFieldDisplayed(2)).toBe(false);
-  });
 });
